Teilchen plazieren und anzeigen Prototype

+++ E-Feld-Simulation/E-Feld-Simulation/Teilchen.h
@@ -5,11 +5,11 @@
 class Teilchen
 {
 public:
-	Teilchen(Physik_Value Ladung, sf::Vector2f position);
+	Teilchen(Physik_Value Ladung, sf::Vector2i position);
 	~Teilchen();
 
 	Physik_Value Q;
-	sf::Vector2f pos;
+	sf::Vector2i pos;
     
 	sf::CircleShape Erstelle_Kreis_Ladung(Physik_Value Radius);
 };
